Introduction Read
After being scammed for 3.5 ETH on UniSwaps through a paired contract trade (where 90% of your tokens were the TXNFee for initiating the trade) I decided to create a token / coin project to help prevent the spread of scam artists stealing good peoples hard earned assets. 

The principal benefits of a Decentralized environment are first and foremost the ease  and speed at which I formation can be spread. However when something becomes too decentralized, there is no longer a single point that can be held accountable.

4 days after the contract pairing had been introduced it was still actively stealing people's assets from UniSwaps with no sign of them interceding. It's very easy to throw your hands up and say not my problem but it's this type of attitude that has led to rampant theft in the emerging DeFi landscape. 

I'm looking to implement a blockchain concept in conjunction with a dedicated central server as a dedicated human verification point to verify the any submitted address/Contract pairings that matches key sets of suspicious transactional data or known code used in various scam contracts. 

In addition we hope to create a broad listing of all traded crypto currency and establish a safety rank for interacting with each of these individual DeFi assets. 

While the specifics of the requirements needed to earn the varying level of approval will vary, in order for us to take accountability for what we produce the final check of having a 2/3 of the established group of project collaborators will be needed via a digital signature. 

Since many scams are occurring Using tokens built of Etherium I'm also considering the future collaboration with other projects that specialize in cross communication between Etherium and other block chains. 

The ultimate goal to create a well established system that can be immediately integrated to existing exchanges and display a safety rating next to each trading pool. 

Until we can achieve such a feat having even a simple contract search that displays any red flags for people to check before engaging in any trades. 

While it will be impossible to 100% secure this Wild West of trading, when people to find themselves scammed we are proposing a reward for being the first to report it. This incentive will both help quickly spread information on the scam as well as alleviate a portion of the scammed persons loss.